requirements.txt
flask==1.0.2
pillow==5.1.0
numpy==1.14
keras==2.1.6
tensorflow==1.8
